services/rider_service/requirements.txt
grpcio
grpcio-tools
protobuf
pika
SQLAlchemy
psycopg2-binary